Fixes date parse logic
