--- src/utis/utils.ts
@@ -1,5 +1,0 @@
-export function getRandomInteger(min: number, max: number) {
-    min = Math.ceil(min);
-    max = Math.floor(max);
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-}
